requirements.txt
flask
python-dotenv
flask-cors
SQLAlchemy
pandas
requests